У точки удален конструктор копирования

+++ amounts.h
@@ -19,11 +19,11 @@ public:
 	// »нициализаци€ класса. ≈го инициализаци€ должна происходить после инициализации массивов в World, когда модель готова к первому тику.
 	void Init();
 
-	unsigned long long getResAmounts(int index) { return arResAmounts ? arResAmounts[index] : 0; }
+	unsigned long long getResAmounts(size_t index) { return arResAmounts ? arResAmounts[index] : 0; }
 
 	// ”величивает количество указанного элемента. ћожно увеличивать на отрицательную величину.
-	void incAmount(int index, unsigned long long amnt) { arResAmounts[index] += amnt; }
-	void decAmount(int index, unsigned long long amnt) { arResAmounts[index] -= amnt; }
+	void incAmount(size_t index, unsigned long long amnt) { arResAmounts[index] += amnt; }
+	void decAmount(size_t index, unsigned long long amnt) { arResAmounts[index] -= amnt; }
 
 private:
